requirements.txt
pymupdf>=1.23.0
requests>=2.31.0
playwright>=1.43.0
python-dotenv